templates-backend/pyproject.toml
[project]
name = "templates-backend"
description = "Backend of the website Template."

version = "1.1.0"

license = "CC-BY-NC-SA-4.0"
authors = [
    { name = "Vincent 'Philaeux' LAMOTTE", email = "vinc.lamotte@gmail.com" }
]

requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.1",
    "fastapi[all]>=0.117.1",
    "psycopg[binary]>=3.2.10",
    "sqlalchemy>=2.0.39",
    "strawberry-graphql>=0.282.0",
    "strawberry-sqlalchemy-mapper>=0.7.0",
    "strawchemy>=0.19.0",
    "uvicorn>=0.34.0",
    "python-jose>=3.3.0",
    "requests>=2.31.0",
    "apscheduler>=3.11.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=2.0.0",
    "sphinxcontrib-mermaid>=0.9.2",
    "enum-tools[sphinx]>=0.12.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.5"
]
